Merge done
All brokerd process are now identified and registered
Ctrl+C well handled by broker : it will ask brokerd to reverse the effects of the command if possible and will return the focus to the user only when the command is done.
When a user try to delete a non-existing resource the associated metadatas will be removed.
Shares creation / mount options are now finer
Remotedesktop xfce_config fixed